Reverted split of elf2arm
