+++ services/elf2arm.cc
@@ -49,7 +49,10 @@
 #include "services/elf2arm.h"
 #include "armProc/blockdev_params.h"
 #include "facilities/arch.h"
+
+#ifndef MKDEV_BUILD
 #include "services/symbol_table.h"
+#endif
 
 /*
  * Functions throughtout this module close over this global!
@@ -66,6 +69,7 @@ static const size_t kBlockSize = 4096;
 
 #define elfError(emsg) ({*error = (char *) emsg; return 1;})
 
+#ifndef MKDEV_BUILD
 struct elfSymbol {
     elfSymbol(const std::string& name, Elf32_Sym* details)
         : name(name), details(details) {}
@@ -99,6 +103,7 @@ private:
     Elf_Data* data;
     Elf32_Sym* current;
 };
+#endif
 
 uint32_t toTargetEndian(uint32_t x);
 
@@ -110,6 +115,7 @@ const char* fileName = NULL;
 // Be verbose
 // static bool verbose = false;
 
+#ifndef MKDEV_BUILD
 static Elf_Scn* getSectionByType(Elf32_Word type, Elf *elf)
 {
     Elf_Scn* sd;
@@ -122,6 +128,7 @@ static Elf_Scn* getSectionByType(Elf32_Word type, Elf *elf)
     }
     return NULL;
 }
+#endif
 
 int initf(const char * fileName, e2adata* data, char **error)
 {
@@ -160,12 +167,12 @@ inline uint32_t toTargetEndian(uint32_t x, Elf32_Ehdr *elfHeader)
 #endif
 }
 
+#ifndef MKDEV_BUILD
 elfSymbolTableIterator::elfSymbolTableIterator(Elf* elf)
     : elf(elf),
       stSec(getSectionByType(SHT_SYMTAB, elf)),
       stSecHeader(elf32_getshdr(stSec))
 {
-    (void)elf;
     current = NULL;
 
     data = NULL;
@@ -215,6 +222,7 @@ elfSymbolTableIterator elfSymbolTableIterator::operator++(int)
     ++(*this);
     return result;
 }
+#endif
 
 static int readCore(e2adata *edata, char **error, coreElf *newelf, uint8_t **dBuf, uint8_t **tBuf, Word *dSize, Word *tSize){
     // Check ELF object type
@@ -241,6 +249,7 @@ static int readCore(e2adata *edata, char **error, coreElf *newelf, uint8_t **dBu
     bool foundTextSeg = false;
     uint8_t* textBuf = NULL;
 
+#ifndef MKDEV_BUILD
     header[AOUT_HE_TAG] = COREFILEID;
 
     //prepare data and text buffers
@@ -293,6 +302,60 @@ static int readCore(e2adata *edata, char **error, coreElf *newelf, uint8_t **dBu
 
     header[AOUT_HE_TEXT_OFFSET] = 0;
     header[AOUT_HE_DATA_OFFSET] = header[AOUT_HE_TEXT_MEMSZ];
+#else
+    header[AOUT_HE_TAG] = AOUTFILEID;
+
+    //prepare data and text buffers
+    for (size_t i = 0; i < phtSize; i++) {
+        if (pht[i].p_type != PT_LOAD)
+            continue;
+        if (pht[i].p_flags == (PF_R | PF_W)) {// data segmet
+            if (foundDataSeg)
+                elfError("Redundant .data program header table entry");
+            foundDataSeg = true;
+            header[AOUT_HE_DATA_MEMSZ] = pht[i].p_memsz;
+            header[AOUT_HE_DATA_VADDR] = pht[i].p_vaddr;
+            header[AOUT_HE_DATA_FILESZ] = (pht[i].p_filesz / kBlockSize) * kBlockSize; //rounds memory size to 4k mem blocks sixe (actual memory to be allocated)
+            if (header[AOUT_HE_DATA_FILESZ] < pht[i].p_filesz)
+                header[AOUT_HE_DATA_FILESZ] += kBlockSize;
+            if (header[AOUT_HE_DATA_FILESZ] > 0) {
+                dataBuf = new uint8_t[header[AOUT_HE_DATA_FILESZ]];
+                *dSize = header[AOUT_HE_DATA_FILESZ];
+                std::fill_n(dataBuf, header[AOUT_HE_DATA_FILESZ], 0);
+            } else {
+                dataBuf = NULL;
+            }
+        } else if (pht[i].p_flags == (PF_R | PF_X)) {// text segment
+            if (foundTextSeg)
+                elfError("Redundant .text program header table entry");
+            if (pht[i].p_memsz == 0)
+                elfError("Empty .text segment");
+            foundTextSeg = true;
+            header[AOUT_HE_TEXT_MEMSZ] = pht[i].p_memsz;
+            header[AOUT_HE_TEXT_VADDR] = pht[i].p_vaddr;
+            header[AOUT_HE_TEXT_FILESZ] = ((AOUT_HE_TEXT_MEMSZ) / kBlockSize) * kBlockSize;
+            if (header[AOUT_HE_TEXT_FILESZ] < header[AOUT_HE_TEXT_MEMSZ])
+                header[AOUT_HE_TEXT_FILESZ] += kBlockSize;
+            textBuf = new uint8_t[header[AOUT_HE_TEXT_FILESZ]];
+            *tSize = header[AOUT_HE_TEXT_FILESZ];
+            std::fill_n(textBuf, header[AOUT_HE_TEXT_FILESZ], 0);
+        } else {
+            fprintf(stderr, "Warning: unknown program header table entry %u\n", (unsigned int) i);
+        }
+    }
+
+    if (!foundTextSeg)
+        elfError("Missing .text program header");
+
+    if (!foundDataSeg) {
+        header[AOUT_HE_DATA_MEMSZ] = header[AOUT_HE_DATA_FILESZ] = 0;
+        header[AOUT_HE_DATA_VADDR] = header[AOUT_HE_TEXT_VADDR] + header[AOUT_HE_TEXT_MEMSZ];
+        dataBuf = NULL;
+    }
+
+    header[AOUT_HE_TEXT_OFFSET] = 0;
+    header[AOUT_HE_DATA_OFFSET] = header[AOUT_HE_TEXT_FILESZ];
+#endif
 
     // Scan sections and copy data to a.out segments
     Elf_Scn* sd;
@@ -331,6 +394,7 @@ static int readCore(e2adata *edata, char **error, coreElf *newelf, uint8_t **dBu
     return 0;
 }
 
+#ifndef MKDEV_BUILD
 static int readStab(e2adata *edata, char **error, Word asid, SymbolTable **rstab)
 {
     static const char* const typeName[] = { "", "OBJ", "FUN" };
@@ -454,6 +518,41 @@ coreElf::coreElf(const char *fname, Word asid){
         }
     }
 }
+#else
+coreElf::coreElf(const char *fname){
+    daddr = taddr = 0;
+    dsize = tsize = 0;
+    error = (char *) "";
+    tread = false;
+    if(!initf(fname, &edata, &error))
+        readCore(&edata, &error, this, &dbuf, &tbuf, &dsize, &tsize);
+}
+
+int coreElf::read(void *dest, unsigned int size, unsigned int count){
+    unsigned int i;
+    uint8_t *tp = (uint8_t *) dest;
+    for(i = 0; i < (size * count); i++, tp++){
+        if(tread){ //read .data
+            if(daddr < dsize){
+                *tp = dbuf[daddr++];
+            } else
+                break;
+        } else { //read .text
+            if(taddr < tsize){
+                *tp = tbuf[taddr++];
+            } else {
+                tread = true;
+                i--;
+                tp--;
+            }
+        }
+    }
+
+    return i;
+
+}
+
+#endif
 
 coreElf::~coreElf(){
     if(dbuf != NULL)
